==> hyesue_jang/3_report_intro_additional_analysis_hj.Rmd <==
---
title: "Group Project Report (draft)"
author: "Hyesue Jang"
date: "12/5/2019"
output: html_document
---

This is a markdown file for the group project write-up. This file includes introduction, additional analysis, and reference sections.
Last modified 12/5/2019

# Introduction

According to the American Time Use Survey (ATUS, https://www.bls.gov/charts/american-time-use/activity-by-work.htm), 85.8 percent of males and 66.5 percent of females work more than 40 hours per week in the U.S. Working overtime has been associated with diverse health problems, especially ischemic heart diseases and cerebrovascular diseases (Nishiyama & Johnson, 1997). Therefore, this project aims to examine whether working overtime is associated with the blood pressure level using the National Health and Nutrition Examination Survey (NHANES) data. In order to produce sufficient estimation sample size, this study use a pooled data from the three two-year cycle data sets: 2011-2012, 2013-2014, and 2015-2016.

It is well known that various biological and environmental factors influence the blood pressure level (Lauer, 1993; Papathanasiou, 2015). To control for the potential influences of such factors, we have systematically reviewed the literature to select the important factors that might affect the abnormal blood pressure levels of the respondents. Final predictors included in the model were gender, age, body mass index (BMI), alcohol consumption level, hours of sleep, and history of smoking. These final predictors closely matched with those from other published studies that observed the association between work hours and blood pressure levels  (Hayashi, 1996; Nakamura et al. 2012).

In terms of defining the criteria for the key variables of interest, working overtime were defined as working more than 40 hours per week. For the additional logistic regression analyses, abnormal blood pressure was defined as having systolic blood pressure higher than 140mm Hg and diastolic blood pressure higher than 90mm Hg following the guide from the American Heart Association. 

In summary, this study will examine whether working overtime (> 40 hours) is associated with the risk of having abnormal systolic and diatolic blood pressure level after controlling for other factors that could affect the risk.


# Additional analysis
### Hyesue Jang

To examine the effects of working overtime on the risk of abnormal blood pressure, I conducted logistic regression analysis. Abnormal blood pressure was defined as having systolic blood pressure higher than 140mm Hg and diastolic blood pressure higher than 90mm Hg following the guide from the American Heart Association. Similar to main analysis, working overtime, gender, age, bim, alcohol consumption level, and smoke history were entered into the model. The model was fit using the glm function in R. The effects of overwork were examined for the risk of abnormal systolic blood pressure and the risk of abnormal diastolic blood pressure, separately. 

The following code shows how logistic regression analysis was run in R: 

```{r eval=FALSE}
#logistic regression with normal vs. abnormal BP outcome

#Systolic blood pressure
logistic_SBP <- glm(SBP_bi ~ overtime + gender + age + bmi +
alcohol_adj + sleep + smoke,
family = binomial(link = 'logit'),
data = df)
summary(logistic_SBP)

#Diastolic blood pressure
logistic_DBP <- glm(DBP_bi ~ overtime + gender + age + bmi +
alcohol_adj + sleep + smoke,
family = binomial(link = 'logit'),
data = df)
summary(logistic_DBP)
```

**Systolic blood pressure** <br/>
 The results suggest that working overtime increase the risk of having abnormal systolic blood pressure level, however, its effect was not significant (&beta; = 0.23, p = 0.43) Among the predictors, age, BMI had significant effects on the risk. Specifically, older age was associated with increased risk of having abnormal systolic blood pressure (&beta; = 0.05, p < 0.001) and higher BMI was associated with having abnomral systolic blood pressure (&beta; = 0.07, p < 0.01).

**Diastolic blood pressure** <br/>
 The results suggest that working overtime increase the risk of having abnormal diastolic blood pressure, however, its effect was not significant (&beta; = 0.41, p = 0.26) Among the included predictors, age, BMI had significant effects on the risk of having abnormal diastolic blood pressure level. Specifically, older age was associated with increased risk of having abnormal diastolic blood pressure level (&beta; = 0.05, p = 0.010) and higher BMI was associated with having abnomral diastolic blood pressure level (&beta; = 0.09, p < 0.01).

# References
Hayashi, T., Kobayashi, Y., Yamaoka, K., & Yano, E. (1996). Effect of overtime work on 24-hour ambulatory blood pressure. Journal of occupational and environmental medicine, 38(10), 1007-1011.

Lauer, R. M., Mahoney, L. T., Clarke, W. R., & Witt, J. (1993). Childhood predictors for high adult blood pressure: the Muscatine Study. Pediatric Clinics of North America, 40(1), 23-40.

Nakamura, K., Sakurai, M., Morikawa, Y., Miura, K., Ishizaki, M., Kido, T., ... & Nakagawa, H. (2012). Overtime work and blood pressure in normotensive Japanese male workers. American journal of hypertension, 25(9), 979-985.

Nishiyama, K., & Johnson, J. V. (1997). Karoshi--death from overwork: occupational health consequences of Japanese production management. International Journal of Health Services, 27(4), 625-641.

Papathanasiou, G., Zerva, E., Zacharis, I., Papandreou, M., Papageorgiou, E., Tzima, C., ... & Evangelou, A. (2015). Association of high blood pressure with body mass index, smoking and physical activity in healthy young adults. The open cardiovascular medicine journal, 9, 5.
